Fix BusyWindowPropagation w/ caching for some corner cases with dependent tasks
Affected configurations:
Dependent task chains with >= 3 tasks on the same resource where one of the
downstream tasks has a higher priority than the first task.

Cause:
The BusyWindowPropagationModel object only stored the reference to the
TaskResult object. The deltamin_func() of the BusyWindowPropagationModel then
got the busy_times from self.task_result.busy_times.
However, the compute_wcrt() on the first task clears the busy_times of its
TaskResult object, so that the deltamin_func() of the second task only returns
the input event model of the first task as there are no busy_times available.

Solution:
When initializing the BusyWindowPropagationModel, the reference to the
busy_times is stored directly.

Remark: It didn't seem to affect the results when caching was disabled.